refactor/tweak: Blob gamemode+event fix and rework (#5312)
* refactor/tweak: blob gamemode+event fix and rework

* fix linter error in blob_storm.dm

* Исправление перевода


* Исправление перевода code/_onclick/hud/blob_overmind.dm


* fix comment in blob_storm.dm

* fix get_blob_candidates in blob.dm

* fix blob.dm

* bugfixes

* fix event mouse min burst time

* fix bugs and tranclate errors + refactor

* incorrect reagent bugfix + colored spans to damage type

* blob reagent exist check

* some spans refactor

* fix

* remove </span> in code/__DEFINES/blob.dm


* blobize null fix in code/datums/mind.dm


* mind mob check + remove spawn

* fix

* Async burst

* remove excess proc

* fix mob exist check in code/datums/mind.dm


* Comments to new variables + spaces to tab replace in proc

---------

+++ code/game/objects/effects/effects.dm
@@ -148,5 +148,5 @@
 		qdel(src)
 
 /obj/effect/decal/blob_act(obj/structure/blob/B)
-	if(B && B.loc == loc)
+	if(B && B.loc == loc && !QDELETED(src))
 		qdel(src)
